Initialize map for env vars
